requirements.txt
certifi==2019.3.9
Django==2.1.7
django-filter==2.1.0
djangorestframework==3.9.2
Markdown==3.1
numpy==1.16.2
pandas==0.24.2
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
